create graph components
