+++ frontend/src/components/utils/Graph.tsx
@@ -1,76 +1,28 @@
-import React, {useState, useEffect} from "react"
-import Paper from '@material-ui/core/Paper';
+import { ArgumentAxis, Chart, SplineSeries, ValueAxis } from "@devexpress/dx-react-chart-material-ui"
+import { ValueScale } from "@devexpress/dx-react-chart"
+import Paper from "@material-ui/core/Paper"
 
-import "date-fns"
-import DateFnsUtils from "@date-io/date-fns"
-
-import {
-  ArgumentAxis,
-  ValueAxis,
-  Chart,
-  SplineSeries,
-  Legend,
-} from "@devexpress/dx-react-chart-material-ui";
-import { ValueScale, Animation } from '@devexpress/dx-react-chart';
-import {
-  MuiPickersUtilsProvider,
-  KeyboardDatePicker
-} from "@material-ui/pickers"
-import { Line } from 'react-chartjs-2';
-
-import { getWeights } from "lib/api/weights"
-import { WeightData } from "interfaces";
-
-
-
-export const Graph = () => {
-  const selectDate = new Date();
-  const [ weights, setWeights ] = useState<WeightData []>([])
-  const [ maxWeight, setMaxWeight ] = useState<string>("")
-  const [ date, setDate ] = useState<Date | null>(selectDate)
-
-
-  const handleChange = (date: Date | null) => {
-    setDate(date);
-  }
-
-  const indexWeights = async () => {
-
-    try {
-      const res = await getWeights ()
-      console.log(res)
-
-      if(res?.status === 200) {
-        setWeights(res.data)
-      }
-    } catch(err) {
-      console.log(err)
-    }
-  }
-
-  // setMaxWeight(weights.reduce((a,b) => a.kg > b.kg?a:b).kg)
-  // const modifyKgDomain = () => [0, parseFloat(maxWeight)+20]
-
-  useEffect(() => {
-    indexWeights();
-  }, []);
+export const Graph = (props: any) => {
+  //受け取ったpropsをChartのデータに割り当てる
+  const { weightsData } = props;
 
   return (
     <>
       <Paper>
-
-        <MuiPickersUtilsProvider utils={DateFnsUtils}>
-          <KeyboardDatePicker
-            label="登録年月日"
-            format="MM/dd/yyyy"
-            id="date-picker-dialog"
-            value={date}
-            onChange={handleChange}
-            KeyboardButtonProps={{
-              "aria-label": "change date"
-            }}
-          />
-          </MuiPickersUtilsProvider>
+        <Chart
+          data={weightsData.map((weight: any) => {
+            const container = {
+              day: new Date(`${weight.date}`).getDate(),
+              daykg: weight.kg
+            }
+            return container
+          })}
+        >
+          <ValueScale name="daykg" modifyDomain={()=>[0,100]}/>
+          <ArgumentAxis />
+          <ValueAxis scaleName="daykg" showTicks/>
+          <SplineSeries valueField="daykg" argumentField="day" scaleName="daykg" />
+        </Chart>
       </Paper>
     </>
   )
